db.Integer takes no parameters

+++ app/models.py
@@ -13,7 +13,7 @@ class Ticket(db.Model):
   id = db.Column(db.Integer, primary_key=True)
   buyer_email = db.Column(db.String(128), index=True, unique=True)
   buyer_tel = db.Column(db.String(16))
-  quanlity = db.Column(db.Integer(16))
+  quanlity = db.Column(db.Integer)
   bought_at = db.Column(db.DateTime())
   amount = db.Column(db.Integer)
